src/main/java/org/dteliukov/security/SecurityPassword.java
package org.dteliukov.security;

import javax.crypto.SecretKeyFactory;
import javax.crypto.spec.PBEKeySpec;
import java.security.NoSuchAlgorithmException;
import java.security.SecureRandom;
import java.security.spec.InvalidKeySpecException;
import java.util.Arrays;
import java.util.Base64;
import java.util.Random;

public class SecurityPassword {
    private static final int ITERATIONS = 10000;
    private static final String SALT = "t6kogcjo2OgnjbyV/SOh5s42TvQoCt+RhsZ5rlqZpMw=";
    private static final int KEY_LENGTH = 256;

    public static byte[] hash(char[] password, byte[] salt) {
        PBEKeySpec spec = new PBEKeySpec(password, salt, ITERATIONS, KEY_LENGTH);
        Arrays.fill(password, Character.MIN_VALUE);
        try {
            SecretKeyFactory skf = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
            return skf.generateSecret(spec).getEncoded();
        } catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
            throw new AssertionError("Error while hashing a password: " + e.getMessage(), e);
        } finally {
            spec.clearPassword();
        }
    }
    public static String generateSecurePassword(String password) {
        byte[] securePassword = hash(password.toCharArray(), SALT.getBytes());

        return Base64.getEncoder().encodeToString(securePassword);
    }

    public static boolean verifyUserPassword(String providedPassword, String securedPassword) {
        String newSecurePassword = generateSecurePassword(providedPassword);

        return newSecurePassword.equalsIgnoreCase(securedPassword);
    }
}
